refactor getTrips from TripsPage and delete token from storage if it is invalid
