+++ src/client/api/trip.js
@@ -1,0 +1,36 @@
+import axios from 'axios';
+
+import APIS from '../constants/apis.js';
+import PATHS from '../constants/paths';
+
+function getTrips (instance) {
+  return axios
+    .request({
+      url: APIS.trips,
+      method: 'get',
+      headers: {
+        token: localStorage.getItem('token'),
+        platform: localStorage.getItem('platform')
+      }
+    })
+    .then(function (response) {
+      console.log(response.data);
+      const { trips } = response.data;
+      instance.setState({ trips });
+      instance.setState({ isDoneFetching: true });
+      instance.setState({ isLoading: false });
+    })
+    .catch(function (error) {
+      if (error.response && error.response.status === 401) {
+        localStorage.removeItem('token');
+        localStorage.removeItem('platform');
+        instance.routeChange(PATHS.login);
+        return;
+      }
+      alert(error.message);
+    });
+}
+
+export default {
+  getTrips
+};
