+++ src/client/constants/apis.js
@@ -4,11 +4,11 @@ const root = 'http://localhost:3001/api/v1';
 const APIS = {
   login: `${root}/login`,
   loginGuest: `${root}/login/guest`,
+  trips: `${root}/trips`,
   trip: (tripId = '') => `${root}/trip/${tripId}`,
   vote: `${root}/trip/vote`,
   voteResults: (tripId) => `${root}/trip/${tripId}/vote/results/`,
   placesToVote: (tripId, userId) => `${root}/trip/${tripId}/vote/user/${userId}`,
-  userTrips: (userId) => `${root}/user/${userId}/trips`,
   place: (placeId) => `${root}/place/${placeId}`,
   tripFriend: (tripId) => `${root}/trip/${tripId}/members`
 };
